Flask_Process_Server/requirements.txt
Flask==3.0.3
numpy==1.26.4
tensorflow-cpu==2.16.1

